Address review feedback: extract shared util, fix over-fetching, add test
- Move isPatternWrapper to @/lib/constraints.ts (shared location per
  project conventions, removes cross-feature import coupling)
- Filter agentIds to only fetch configs for agents with
  source_action_configuration_id (avoids unnecessary API calls)
- Add test for missing config scenario (config ID set but not in map)
