+++ frontend/src/lib/constraints.ts
@@ -1,0 +1,9 @@
+/** Check if a stored parameter value is a $pattern wrapper object. */
+export function isPatternWrapper(value: unknown): value is { $pattern: string } {
+  return (
+    typeof value === "object" &&
+    value !== null &&
+    "$pattern" in value &&
+    typeof (value as Record<string, unknown>).$pattern === "string"
+  );
+}
